feat: updated test cases

+++ src/types/credits-admin-types.ts
@@ -1394,7 +1394,7 @@ export type Customer = Node & {
   lastName: Scalars['String']['output'];
   orders: OrderList;
   phoneNumber?: Maybe<Scalars['String']['output']>;
-  storeCredit: Scalars['Int']['output'];
+  storeCredit: Scalars['Float']['output'];
   title?: Maybe<Scalars['String']['output']>;
   updatedAt: Scalars['DateTime']['output'];
   user?: Maybe<User>;
@@ -5662,7 +5662,7 @@ export type Seller = Node & {
   customFields?: Maybe<Scalars['JSON']['output']>;
   id: Scalars['ID']['output'];
   name: Scalars['String']['output'];
-  storeCredit: Scalars['Int']['output'];
+  storeCredit: Scalars['Float']['output'];
   updatedAt: Scalars['DateTime']['output'];
 };
